% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observer.R
\docType{data}
\name{observer}
\alias{observer}
\title{Observer}
\format{A data.frame with the following columns:
  
  \code{observer_id}: Unique integer id of observer
  
  \code{first_name}: Full first name of observer
  
  \code{last_name}: Full last name of observer
  
  \code{birth_date}: Date of birth (YYYY-mm-dd)
  
  \code{email}: Full email address of observer}
\usage{
observer
}
\description{
Contains observer records including an observer id, name, and other metadata
}
\keyword{datasets}

