% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sightings_really_bad.R
\docType{data}
\name{sightings_really_bad}
\alias{sightings_really_bad}
\title{Sightings - Really Bad Data Management}
\format{A data.frame with the following columns:
  
  \code{Observer}: Initials of observer - Note inconsistency of
  capitalization. Furthermore, initials may not always be unique
  
  \code{Location}: Title or Initials of Location - Note differences of
  records.
  
  \code{Lat}: Latitude
  
  \code{Lon}: Longitude
  
  \code{SPECIES_1}: Species Abbreviation - Why is SPECIES in all caps?
  
  \code{Date_1}: Date without time
  
  \code{Time_1}: Time of observation - Note that some have another date from
  1899
  
  \code{How recognized?_1}: What technique was used to recgonize the bird - 
  Why is there a question mark in the column name?
  
  \code{SPECIES_2}: Species of second observation of same observer in
  location - poor structure
  
  \code{Date_2}: Date of second observation of same observer in location -
  poor structure
  
  \code{Time_2}: Time of second observation of same observer in location -
  poor structure
  
  \code{How recognized?_2}: Recognition type of second observation of same
  observer in location - poor structure
  
  \code{SPECIES_3}: Species of third observation of same observer in location
  - painful
  
  \code{Date_3}: Date of third observation os same ovserver in location -
  painful
  
  \code{Time_3}: Time of third observation of same observer in location -
  painful
  
  \code{How recognized?_3}: Recognition type of third observation of same
  observer in location - painful
  
  \code{SPECIES_4}: Species of fourth observation of same observer in
  location - if exists
  
  \code{Date_4}: Date of fourth observation of same observer in location - if
  exists
  
  \code{Time_4}: Time of fourth observation of same observer in location - if
  exists
  
  \code{How recognized?_4}: Recognition type of fourth observation of same
  observer in location - if exists
  
  \code{comments}: Miscellaneous comments - Which observation does this refer
  to? 1,2,3,4?}
\usage{
sightings_really_bad
}
\description{
Contains bird observations with really bad formatting
}
\keyword{datasets}

