% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_bad.R
\docType{data}
\name{species_bad}
\alias{species_bad}
\title{Species - Bad Data Management}
\format{A data.frame with the following columns:
  
  \code{ID}: Unique integer ID of bird species - Note is an integer ID needed
  if the species abbreviation is unique?
  
  \code{species Abbreviation}: Four letter abbreviation of species - Note
  space in column name
  
  \code{genus}: Genus of bird
  
  \code{species}: Full name of species of bird
  
  \code{common name}: Full common name of species - Note space in column name}
\usage{
species_bad
}
\description{
Contains species records with poor data management
}
\keyword{datasets}
