% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\docType{data}
\name{species}
\alias{species}
\title{Species}
\format{A data.frame with the following columns:
  
  \code{species_id}: Unique four letter abbreviation of species
  
  \code{genus}: Genus of bird
  
  \code{species}: Full name of species of bird
  
  \code{common_name}: Full common name of species}
\usage{
species
}
\description{
Contains species records including a species id, genus, species, and common
name
}
\keyword{datasets}
